make sure we get UTF-8 encoded localized strings
When using locales like de_DE month names like März will be encoded in latin1 encoding while the webpage is in UTF-8 encoding. To get this in line, we append .UTF-8 if there is no '.' in the locale, since a '.' would indicate that some other encoding information is already present in the string.