ENHANCEMENT: Set SS locale to match requested one
Fluent is great for changing CMS edited content to a different locale per request.
However this does not alter all the template and PHP managed translation such as
FormField labels, or template controlled headings. This leads to odd end user
experiences where someone may be viewing the German content, and have it labelled
by English. This fix sets the i18n class locale to the same as the requested one
used by Fluent to give a consistent and far less confusing experience. This is
limited to the 'front end' (non CMS) site as e.g. it is feasible that a content
editor in Germany who is fluent in English would want the interface to remain in
German.