ENHANCEMENT Support fulltext search per locale
In order to be able to search content, we must first be able to add it
to the index. To add the content in each locale (which can vary) we must
support a Variant, much like that of the `Versioned` module. Fluent
itself takes care of modifying the query to select content, we just must
allow the Reindex task to attain and adjust it's settings.