Update composer requirement constraints
Although inconsequential, the current constraints are all marked as dev - an overhang from when SilverStripe 4 was in pre-release stages. This is no longer necessary, so we can remove them in order to keep the constraints more in line with other modules (such as the requirements here) and easier to read.